# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FlyRank Auth API",
    "version": "1.0.0",
    "description": "Secure API with Supabase Auth for signup, login, logout, and protected routes."
  },
  "servers": [{ "url": "http://localhost:3000" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "Create a new user account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created" },
          "400": { "description": "Bad request" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Authenticate user & return JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful" },
          "400": { "description": "Bad request" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Terminate the user session",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "Logged out" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/public/info": {
      "get": {
        "summary": "Read public, unprotected data",
        "responses": {
          "200": { "description": "Public info" }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Read private user profile data",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Profile data" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "summary": "Read private dashboard data",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Dashboard data" },
          "401": { "description": "Unauthorized" }
        }
      }
    }
  }
}